Add directory for Data and Data Analysis
Added 2 datasets with codes used to analyze the data and all of the figures and results (including the results from a commercial device). Outside this folder is the latest version of the code used to analyze the data (every data set additionally comes with the code that was used for generating the results).